refactor(analytics): Flight log merged into Analytics — one stats surface [M]
The de-dupe wasn't the ask — the collapse was. FlightLog.jsx deleted;
Analytics Overview leads with the profile hero (rally/best/lifetime/
points-yr/best day/today) and ends with Achievements. Trend icon, desktop
sidebar, and More all open the same AnalyticsModal. Verified by headless
render (hero + 4 tabs + badges on one surface).
