+++ CLAUDE.md
@@ -1297,7 +1297,7 @@ Tracked in [GitHub Issues](https://github.com/ryakel/boomerang/issues). Key item
 - **TaskActionsContext:** All task callbacks (`onComplete`, `onSnooze`, `onEdit`, `onExtend`, `onStatusChange`, `onUpdate`, `onDelete`, `onGmailApprove`, `onGmailDismiss`) plus `isDesktop` live in `src/contexts/TaskActionsContext.jsx`. TaskCard receives only `task`, `expanded`, and `onToggleExpand` as props. KanbanBoard and ProjectsView consume actions from context.
 - **Desktop keyboard shortcuts:** `useKeyboardShortcuts` hook — `n` (new), `/` (search), `j`/`k` (navigate), `Enter`/`e` (edit), `x` (complete), `s` (snooze), `Escape` (close), `?` (help). Stack-aware Escape closes topmost modal.
 - **Analytics dashboard:** `GET /api/analytics/history?days=N` returns aggregated completion data (daily counts, by-tag, by-energy, by-size, by-DOW). Client renders daily bar chart, day-of-week patterns, tag/energy/size breakdowns, 52-week GitHub-style heat map (UTC-aligned to match server bucketing), and collapsible completed task search. Pure CSS charts, no charting libraries.
-- **Achievements (`src/badges.js` + `src/components/BadgesGrid.{jsx,css}`):** locally-derived badges with DURABLE earned state (`settings.badges_earned`, key-union guarded — deleting rows never un-earns). Each badge has a `tier` (bronze/silver/gold) that drives the card tint; the grid header carries a tier legend so the colors read. Cards are equal-height (`grid-auto-rows: 1fr`) with bottom-pinned footers so earned dates / progress bars align. Tapping a card opens a detail overlay (`BadgeDetail`): earned date or progress + "N to go", plus a done/outstanding **checklist** for set-shaped badges (e.g. Balanced Diet → the six energy types with this-week checks, via `checklist`/`checklistTitle` on the badge). Shared by AnalyticsModal + the Kept Flight log.
+- **Achievements (`src/badges.js` + `src/components/BadgesGrid.{jsx,css}`):** locally-derived badges with DURABLE earned state (`settings.badges_earned`, key-union guarded — deleting rows never un-earns). Each badge has a `tier` (bronze/silver/gold) that drives the card tint; the grid header carries a tier legend so the colors read. Cards are equal-height (`grid-auto-rows: 1fr`) with bottom-pinned footers so earned dates / progress bars align. Tapping a card opens a detail overlay (`BadgeDetail`): earned date or progress + "N to go", plus a done/outstanding **checklist** for set-shaped badges (e.g. Balanced Diet → the six energy types with this-week checks, via `checklist`/`checklistTitle` on the badge). Rendered by AnalyticsModal's Overview tab (2026-07-17: the separate Kept Flight log surface was merged INTO Analytics — one stats surface; `src/kept/FlightLog.jsx` deleted, trend icon + sidebar + More all open AnalyticsModal).
 - **AI-assisted search (Done + Activity Log):** Both modals have a search bar with instant local substring filter + debounced AI semantic search via `POST /api/search/ai`. Uses Claude Haiku for cost efficiency. Falls back to LIKE search (done) or local filter (activity) without an API key. Server endpoint accepts `{ query, scope, items? }` — `scope: 'done'` queries DB + AI ranking, `scope: 'activity'` ranks provided items.
 - **Claude response parsing (`claudeText`/`NO_THINKING` in `aiModels.js`, 2026-07-17, born of the Polish-button crash):** Claude Sonnet 5 runs ADAPTIVE THINKING BY DEFAULT when the request omits the `thinking` param — responses can lead with a `thinking` block (empty text under the default display), so `data.content[0].text` is undefined: client sites crashed (`undefined is not an object ... .match`), `?.`-chained server sites silently returned `''` and degraded (Gmail classify, growth-area rephrasing, AI nudges, pattern/tag scans). NEVER read `content[0].text` — use `claudeText(data)` (collects all text blocks) from `aiModels.js`, which both bundles import. Cheap utility calls (inference, classification, one-liners) also spread `...NO_THINKING` (`thinking: {type:'disabled'}`) into the request body — restores pre-Sonnet-5 cost/latency and stops thinking eating small `max_tokens` budgets (the 100-token nudge calls) from the inside. Sonnet 5 also REJECTS non-default sampling params with a 400 — the Gmail classifier's `temperature: 0` had silently killed classification entirely; never send `temperature`/`top_p`/`top_k` to Sonnet 5+. Deliberately left on default adaptive thinking: the Quokka adviser loop (appends `response.content` wholesale, so thinking blocks echo back correctly) and the `research_task`/ladder-draft tools (already filter text blocks; reasoning helps there).
 - **Multi-provider AI + usage dashboard (2026-07-17, migration 043):** the utility-AI surfaces route through two TIERS instead of hardcoded models — `workhorse` (classification, inference, polish, scans, Gmail classify, nudges) and `quick` (one-liners, AI search, push nudges) — resolved by `resolveTierModel(tier, settings)` in `aiModels.js` from `settings.ai_model_workhorse`/`ai_model_quick` (catalog id, or `provider:model-id` for anything newer; defaults Sonnet 5 / Haiku 4.5). **`server/aiGateway.js`** is the single door: `aiComplete({tier, system, user, maxTokens, feature})` routes to Anthropic or OpenAI (chat completions, `max_completion_tokens`, `reasoning_effort:'low'` on gpt-5-family so reasoning doesn't eat the budget — the OpenAI twin of the Sonnet 5 thinking trap) and logs every call into the `ai_usage` table (`logAiUsage`/`getAiUsageSummary` in db.js; cost estimated at insert from `MODEL_CATALOG` pricing, longest-prefix matched because providers echo DATED model ids). Client utility calls go through `POST /api/ai/complete` (headers `x-anthropic-key`/`x-openai-key`, fall back to settings/env `OPENAI_API_KEY`); usage dashboard at `GET /api/ai/usage` renders as Analytics → AI tab (est. cost / calls / tokens + per-provider/model/feature tables). **Pinned Anthropic, NOT tier-routed:** Quokka's agent loop (tool-use shaped), `research_task`/ladder-draft tools, and the vision surfaces (`/api/messages` — attachment OCR `_feature:'ocr'`, research `_feature:'research'`); all of these still LOG usage at their own call sites (feature `quokka` etc.). Modules gate AI features on `aiConfigured(tier)` — never on an Anthropic key directly, or an OpenAI-only setup silently loses every feature. OpenAI key: Settings → Integrations → OpenAI (probe = free GET /v1/models, also in `check_integrations`); model pickers: Settings → Tasks → AI models. `openai_api_key` is IN the Quokka secret blocklist.
