+++ src/AppV2.jsx
@@ -21,7 +21,6 @@ import ProjectsView from './components/ProjectsView'
 import DoneList from './components/DoneList'
 import ActivityLog from './components/ActivityLog'
 import NotificationsModal from './components/NotificationsModal'
-import FlightLog from './kept/FlightLog'
 import RoutinesModal from './components/RoutinesModal'
 import KeptShell from './kept/KeptShell'
 import KeptDesktop from './kept/KeptDesktop'
@@ -108,7 +107,6 @@ export default function AppV2() {
   const [showDone, setShowDone] = useState(false)
   const [showActivityLog, setShowActivityLog] = useState(false)
   const [showNotifications, setShowNotifications] = useState(false)
-  const [showFlightLog, setShowFlightLog] = useState(false)
   const [showMarkdownImport, setShowMarkdownImport] = useState(false)
   const [updateVersion, setUpdateVersion] = useState(null)
   // SpacesHub is the destination for the Spaces tab. Opens a picker
@@ -613,7 +611,6 @@ export default function AppV2() {
   if (showDone) activeModals.push('done')
   if (showActivityLog) activeModals.push('activitylog')
   if (showNotifications) activeModals.push('notifications')
-  if (showFlightLog) activeModals.push('flightlog')
   if (showRoutines) activeModals.push('routines')
   if (showPackages) activeModals.push('packages')
   if (showAdviser) activeModals.push('adviser')
@@ -635,7 +632,6 @@ export default function AppV2() {
     if (showDone) { setShowDone(false); return }
     if (showActivityLog) { setShowActivityLog(false); return }
     if (showNotifications) { setShowNotifications(false); return }
-    if (showFlightLog) { setShowFlightLog(false); return }
     if (showRoutines) { setShowRoutines(false); return }
     if (showPackages) { setShowPackages(false); return }
     if (showAdviser) { setShowAdviser(false); return }
@@ -1465,7 +1461,6 @@ export default function AppV2() {
           onOpenDone={() => setShowDone(true)}
           onOpenActivity={() => setShowActivityLog(true)}
           onOpenNotifications={() => setShowNotifications(true)}
-          onOpenFlightLog={() => setShowFlightLog(true)}
           onOpenSuggestions={() => setShowSuggestions(true)}
           onOpenGrowthAreas={() => setShowGrowthAreas(true)}
           syncStatus={syncStatus}
@@ -1524,7 +1519,6 @@ export default function AppV2() {
           onOpenDone={() => setShowDone(true)}
           onOpenActivity={() => setShowActivityLog(true)}
           onOpenNotifications={() => setShowNotifications(true)}
-          onOpenFlightLog={() => setShowFlightLog(true)}
           onStatusChange={handleStatusChange}
           onOpenSuggestions={() => setShowSuggestions(true)}
           onOpenGrowthAreas={() => setShowGrowthAreas(true)}
@@ -1708,16 +1702,7 @@ export default function AppV2() {
         onClose={() => setShowActivityLog(false)}
         onRestore={handleRestore}
       />
-      <FlightLog
-        open={showFlightLog}
-        onClose={() => setShowFlightLog(false)}
-        onOpenAnalytics={() => { setShowFlightLog(false); setShowAnalytics(true) }}
-        tasks={tasks}
-        routines={routines}
-        records={records}
-        streak={streak}
-        dailyStats={dailyStats}
-      />
+
       <NotificationsModal
         open={showNotifications}
         onClose={() => setShowNotifications(false)}
@@ -1812,6 +1797,7 @@ export default function AppV2() {
         routines={routines}
         records={records}
         streak={streak}
+        dailyStats={dailyStats}
       />
 
       {isDesktop && !isKept && (
