+++ wiki/Version-History.md
@@ -6,6 +6,9 @@ Commit-level changelog for Boomerang, grouped by date. Sizes: `[XS]` trivial, `[
 
 ## 2026-07-17
 
+- refactor(analytics): Flight log merged INTO Analytics — one stats surface [M]
+  - User's actual point, correctly this time: "tell me how analytics is not flight log with a different name... maybe we should collapse them." The previous de-dupe still left two surfaces showing the same stats (stat cards + year ribbon + badges vs summary + year heatmap + charts). Collapsed: `src/kept/FlightLog.jsx` deleted; AnalyticsModal's Overview tab now leads with the profile hero (rally / best rally / lifetime / points-yr / best day / today, reusing the `bm-fl-stats` cards) and ends with Achievements. Every entry point — Kept header trend icon, desktop sidebar (Flight log row removed), More → Analytics — opens the same modal. The DensityRibbon year view retired with the page (the heatmap is the year view; the component survives for LoopsView). Verified by headless render: avatar tap → merged Analytics with hero, 4 tabs, badges.
+
 - fix(analytics): heatmap actually shows data on mobile; AI tab findable; Flight log \u2194 Analytics de-duped [M]
   - **Heatmap ("GitHub style graph is completely broken"):** verified by headless render — the 687px grid sat in a ~308px scroller pinned to the LEFT, so phones showed only the year-old empty half while every active cell was off-screen right; worse, the month labels lived OUTSIDE the scroller and compressed to viewport width, so labels didn't even correspond to columns. Fix: labels moved inside a `heatmap-inner` container at the grid's natural width (labels + columns scroll together), and the scroller auto-positions to NOW on open — recent weeks visible first, swipe back in time. Confirmed post-fix via DOM probe (scrollLeft pinned right, hot cells on-screen) + screenshot.
   - **AI tab ("where is the AI tab???"):** it rendered, but as the third identical segmented-pill row below the range and metric pickers — invisible as navigation. Tabs (Overview/Tasks/Habits/AI) now sit FIRST in the modal, styled as underline tabs (ember active underline, mirrors Settings); the metric toggle + summary card only render on the tabs they apply to (Overview/Tasks).
